requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0.1
django-cors-middleware==1.5.0
django-extensions==3.1.5
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
typing_extensions==4.0.1
